pantone-scraper/requirements.txt
# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
aiohttp>=3.9.4
requests>=2.31.0
PyYAML>=6.0.1
lxml>=4.9.0
